+++ ProiectLicenta/Controllers/PlaylistController.cs
@@ -48,9 +48,13 @@ namespace ProiectLicenta.Controllers
         [Authorize()]
         public virtual async Task<IActionResult> Delete(int id)
         {
-            var obj = GetById(id);
-            await _repository.Delete(id);
-            return Ok(obj);
+            var obj = await _repository.Get(id);
+            if (obj != null)
+            {
+                await _repository.Delete(id);
+                return Ok(obj);
+            }
+            return BadRequest("Item doesn't exist");
         }
     }
 }
