+++ ProiectLicenta/Controllers/SongController.cs
@@ -20,12 +20,12 @@ namespace ProiectLicenta.Controllers
         protected MapperConfiguration configuration;
         Mapper mapper;
         private readonly SongRepository _repository;
-        private readonly UserManager<AppUser> _userManager;
+        private readonly MessageRepository _messageRepository;
 
-        public SongController(SongRepository repository, UserManager<AppUser> repo):base(repository)
+        public SongController(SongRepository repository, MessageRepository messageRepository):base(repository)
         {
             this._repository = repository;
-            this._userManager = repo;
+            this._messageRepository = messageRepository;
             configuration = new MapperConfiguration(cfg =>
             {
                 cfg.CreateMap<SongCreateDTO, Song>().ReverseMap();
@@ -38,6 +38,13 @@ namespace ProiectLicenta.Controllers
         {
             var obj = await _repository.GetByName(name);
             return Ok(obj);
+        }
+        [HttpGet("includes/{id}")]
+        public async Task<IActionResult> GetByIdWithIncludes(int id)
+        {
+            var obj = await _repository.GetSongWithIncludes(id);
+            var objToSend = mapper.Map<SongCreateDTO>(obj);
+            return Ok(objToSend);
         }
         [HttpGet("album/{id}")]
         public IActionResult GetSongsFromAlbumId(int id)
@@ -52,43 +59,31 @@ namespace ProiectLicenta.Controllers
             var songs = _repository.GetAllQuerry().Where(q => q.ArtistId == id);
             return Ok(songs);
         }
-
-        [HttpPost("upload")]
-        public async Task<IActionResult> UploadAudioFile(IFormFile file)
+        [HttpGet("trending/{start}/{cantity}")]
+        public async Task<IActionResult> GetTrending(int start, int cantity)
         {
-            /* 
-             * the content types of Wav are many
-             * audio/wave
-             * audio/wav
-             * audio/x-wav
-             * audio/x-pn-wav
-             * see "https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types"
-            */
-            if (!file.ContentType.Contains("audio"))
-            {
-                return BadRequest("Wrong file type");
-            }
-           // var uploads = Path.Combine(HostingEnvironment.WebRootPath, "uploads");//uploads where you want to save data inside wwwroot
-          //  var filePath = Path.Combine(uploads, file.FileName);
-            var filePath = @"e:\VS_Projects\"+file.Name;
-            using (var fileStream = new FileStream(filePath, FileMode.Create))
-            {
-                await file.CopyToAsync(fileStream);
-            }
-            return Ok("File uploaded successfully");
+            var songs =await _repository.GetTrending(start, cantity);
+            return Ok(songs);
         }
-        
+
         [HttpPost]
-        [Authorize(Roles =UserRoles.Artist + "," + UserRoles.Admin)]
+       // [Authorize(Roles =UserRoles.Artist)]
         public virtual async Task<IActionResult> Create(SongCreateDTO obj)
         {
-            var result = mapper.Map<Song>(obj);
-            result.ImagePath = "";
-            await _repository.Add(result);
-            return Ok(obj);
+            var song = new Song();
+            song.Name = obj.Name;
+            song.Duration = 0;
+            song.ServerLink = "";
+            song.ImagePath = "";
+            song.ArtistId = obj.ArtistId;
+            song.AlbumId = obj.AlbumId;
+            song.GenreId= obj.GenreId;
+            
+            await _repository.Add(song);
+            song = await _repository.GetByName(song.Name);
+            return Ok(song);
         }
         
-
         [HttpPut("update")]
         [Authorize(Roles = UserRoles.Artist + "," + UserRoles.Admin)]
         public virtual async Task<IActionResult> Update(SongCreateDTO obj)
@@ -113,9 +108,59 @@ namespace ProiectLicenta.Controllers
         [Authorize(Roles = UserRoles.Artist + "," + UserRoles.Admin)]
         public virtual async Task<IActionResult> Delete(int id)
         {
-            var obj = GetById(id);
-            await _repository.Delete(id);
-            return Ok(obj);
+            var obj = await _repository.GetSongWithIncludes(id);
+            if (obj != null)
+            {
+                await _repository.Delete(id);
+                return Ok(obj);
+            }
+            return BadRequest("Item doesn't exist");
+        }
+        [HttpPost("addMessage")]
+        //[Authorize()]
+        public async Task<IActionResult> AddMessage(MessageCreateDTO message)
+        {
+            var song = await _repository.GetSongWithIncludes(message.SongId);
+            if (song != null)
+            {
+                Message actual = new Message();
+                actual.Text= message.Text;
+                actual.WhoSentId= message.WhoSentId;
+                actual.SongId= message.SongId;
+                actual.Song = song;
+                await _messageRepository.Add(actual);
+
+                actual = await _messageRepository.Get(actual.Id);
+                song.Messages.Add(actual);
+                return Ok(actual);
+            }
+            return BadRequest("Wrong song id");
+        }
+        [HttpDelete("removeMessage/{songId}/{id}")]
+        //[Authorize()]
+        public async Task<IActionResult> DeleteMessage(int songId,int id)
+        {
+            var song = await _repository.GetSongWithIncludes(songId);
+            var message = await _messageRepository.GetByIdWithIncludes(id);
+            if (song != null)
+            {
+                if(message !=null)
+                {
+                    message.LikesFromUsers.Clear();
+                    await _messageRepository.Update(message);
+                    message = await _messageRepository.GetByIdWithIncludes(id);
+
+                    song.Messages.Remove(message);
+                    await _repository.Update(song);
+
+                    
+                    await _messageRepository.Delete(id);
+                    return Ok(message);
+                }
+                return BadRequest("Wrong message id");      
+            }
+            return BadRequest("Wrong song id");
         }
     }
 }
+
